// src/types/hostBonus/hostBonus.ts
export interface ISelectItemsEarnUpHostBonus {
  img: string
  headerTitle: string
  mainTitle: string
  content: string[]
  footer: string
  footer2: string
  footer3: string
}
export interface ISelectItemsSecureHostBonus {
  headerTitle: string
  content: string
  nameButtonSharecar: string
  footer: string
  footer2: string
  nameButtonAddCar: string
}
export interface ISelectItemsHostDealHostBonus {
  img: string
  title: string
  content: string
}